requirements.txt
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.2.0
tqdm>=4.65.0
pyyaml>=6.0.0
pytest>=7.0.0

